Add grpc client samples and refactor stockqoute sample
