+++ samples/grpc-helloworld/helloworld-service/src/main/java/org/wso2/msf4j/grpc/helloworld/HelloService.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright (c) 2017, WSO2 Inc. (http://wso2.com) All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.wso2.msf4j.grpc.helloworld;
+
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
+
+/**
+ * Hello service resource class.
+ */
+@Path("/hello")
+public class HelloService {
+
+    @GET
+    @Path("/{name}")
+    public String hello(@PathParam("name") String name) {
+        System.out.println("Hello");
+        return "Hello " + name;
+    }
+
+    // This method is needed, If you are using server as gRPC.
+    private Class getProtoService() {
+        return HelloServiceProto.class;
+    }
+}
