+++ samples/grpc-helloworld/helloworld-service/src/main/proto/helloworld.proto
@@ -1,0 +1,37 @@
+// Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+syntax = "proto3";
+
+option java_multiple_files = true;
+option java_package = "org.wso2.msf4j.grpc.helloworld";
+option java_outer_classname = "HelloServiceProto";
+option objc_class_prefix = "HLW";
+
+package org.wso2.msf4j.grpc.helloworld;
+
+// The greeting service definition.
+service HelloService {
+    // Sends a greeting
+    rpc hello (HelloRequest) returns (HelloResponse) {}
+}
+
+// The request message containing the user's name.
+message HelloRequest {
+    string name = 1;
+}
+
+// The response message containing the greetings
+message HelloResponse {
+    string message = 1;
+}
