+++ samples/grpc-stockquote/stockquote-client/pom.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <groupId>org.wso2.msf4j</groupId>
+        <artifactId>msf4j-service</artifactId>
+        <version>2.4.3-SNAPSHOT</version>
+        <relativePath>../../../poms/msf4j-service/pom.xml</relativePath>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>stockquote-client</artifactId>
+
+    <properties>
+        <microservice.mainClass>org.wso2.msf4j.grpc.stockquote.StockQuoteClient</microservice.mainClass>
+    </properties>
+</project>
