+++ samples/grpc-stockquote/stockquote-service/src/main/proto/stockquote.proto
@@ -1,0 +1,47 @@
+// Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+syntax = "proto3";
+
+option java_multiple_files = true;
+option java_package = "org.wso2.msf4j.grpc.stockquote";
+option java_outer_classname = "StockQuoteServiceProto";
+option objc_class_prefix = "HLW";
+
+package org.wso2.msf4j.grpc.stockquote;
+
+// The stockquote service definition.
+service StockQuoteService {
+    rpc getQuote (StockRequest) returns (StockQuote) {}
+    rpc addStock (StockQuote) returns (Empty) {}
+}
+
+message StockQuote {
+    string symbol = 1;
+    string name = 2;
+    double last = 3;
+    double low = 4;
+    double high = 5;
+
+}
+
+message StockRequest {
+    string name = 1;
+}
+
+message StockQuotes {
+    StockQuote stock = 1;
+}
+
+message Empty {
+}
